Kitchen Service/requirements.txt
boto3==1.24.34
botocore==1.27.34
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-pubsub==2.13.4
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
jmespath==1.0.1
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
rsa==4.8
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.10
